+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 # Changelog
 All notable changes to this project will be documented in this file.
 
+## [0.1.5] - 2020-04-09
+### Changed
+- Infinite loader to explicit `Load more` button
+
+### Fixed
+- Service worker update Toast now being shown
+
 ## [0.1.2] - 2020-04-09
 ### Added
 - App Icon added